Proxy: Set production-ready read/write timeouts
The proxy was using the default values when serving requests,
which is infinite. This means if any client hangs for whatever
reason while reading/writing - although this problem mostly happens
during a write operation - it blocks every other process that
is waiting to have a free channel.

With the new adjustments, the server times out a request whenever
it takes longer than expected. These values were adjusted to 5s
for readings, and 10s for writings.